+++ phpunit.xml.dist
@@ -14,8 +14,11 @@
          verbose="true"
 >
     <testsuites>
-        <testsuite name="Package Test Suite">
-            <directory suffix="Test.php">./tests</directory>
+        <testsuite name="Unit">
+            <directory suffix="Test.php">./tests/Unit</directory>
+        </testsuite>
+        <testsuite name="Feature">
+            <directory suffix="Test.php">./tests/Feature</directory>
         </testsuite>
     </testsuites>
 </phpunit>
