+++ tests/Feature/ClientPrivateKeyBasicCredentialsTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Illuminate\Tests\Notifications;
+namespace Illuminate\Notifications\Tests\Feature;
 
 use Vonage\Client;
 use Vonage\Client\Credentials\Basic;
 use Vonage\Client\Credentials\Container;
 use Vonage\Client\Credentials\Keypair;
 
-class ClientPrivateKeyBasicCredentialsTest extends AbstractTestCase
+class ClientPrivateKeyBasicCredentialsTest extends FeatureTestCase
 {
     protected function getEnvironmentSetUp($app)
     {
@@ -19,14 +19,12 @@ class ClientPrivateKeyBasicCredentialsTest extends AbstractTestCase
 
     public function testClientCreatedWithPrivateKeyBasicCredentials()
     {
-        $client = app(Client::class);
+        $credentials = $this->app->make(Client::class)->getCredentials();
 
-        $credentialsObject = $this->getClassProperty(Client::class, 'credentials', $client);
-        $credentialsArray = $this->getClassProperty(Container::class, 'credentials', $credentialsObject);
-        $keypairCredentials = $this->getClassProperty(Keypair::class, 'credentials', $credentialsArray[Keypair::class]);
-        $basicCredentials = $this->getClassProperty(Basic::class, 'credentials', $credentialsArray[Basic::class]);
+        $keypairCredentials = $credentials->asArray()[Keypair::class]->asArray();
+        $basicCredentials = $credentials->asArray()[Basic::class]->asArray();
 
-        $this->assertInstanceOf(Container::class, $credentialsObject);
+        $this->assertInstanceOf(Container::class, $credentials);
         $this->assertEquals(['key' => '===FAKE-KEY===', 'application' => 'application-id-123'], $keypairCredentials);
         $this->assertEquals(['api_key' => 'my_api_key', 'api_secret' => 'my_secret'], $basicCredentials);
     }
