--- tests/TestCase.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Illuminate\Tests\Notifications;
-
-use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
-use PHPUnit\Framework\TestCase as BaseTestCase;
-
-abstract class TestCase extends BaseTestCase
-{
-    use MockeryPHPUnitIntegration;
-}
