+++ tests/Unit/Channels/VonageShortcodeChannelTest.php
@@ -1,16 +1,19 @@
 <?php
 
-namespace Illuminate\Tests\Notifications\Channels;
+namespace Illuminate\Notifications\Tests\Unit\Channels;
 
 use Illuminate\Notifications\Channels\VonageShortcodeChannel;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Notifications\Notification;
-use Illuminate\Tests\Notifications\TestCase;
 use Mockery as m;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
+use PHPUnit\Framework\TestCase;
 use Vonage\Message\Client;
 
 class VonageShortcodeChannelTest extends TestCase
 {
+    use MockeryPHPUnitIntegration;
+
     public function testShortcodeIsSentViaVonage()
     {
         $notification = new NotificationVonageShortcodeChannelTestNotification;
